fix: 하드코딩 되어있던 groupId 수정
